Started back on the project again! Finally have some free time, which is wonderful.
I managed to implement the extremely minimal and barebones for the
keyboard driver! It outputs characters to the screen based on input, and
even has simple backspace (although not for multi-line yet). As well,
there is an issue where the shift key does convert to upper case, but I
can't manage to unset the flag when it is released. I'll get to that
tomorrow probably.