I think I finally see it. I am further refactoring the keyboard, and enhancing it as well. Began implementation of scan table mapping to character coding. I can see it now, of how I will plan out the shell.
For one, there will be a 'shell' module, which will register (in the
mean time at least) with the keyboard HAL ('kbd_hal') for input events,
and use the VGA driver for displaying input. This way the shell has full
control over if it should ignore characters, how the display will
changed based on commands, etc. Sounds pretty fun.